+++ NEWS.md
@@ -25,6 +25,8 @@
 
 * tidyr now requires dplyr >=1.1.0 (#1568, @catalamarti).
 
+* `unnest_longer()` now outputs indices before values (#1486, @andrewshibata)
+
 # tidyr 1.3.1
 
 * `pivot_wider` now uses `.by` and `|>` syntax for the dplyr helper message to